Register type to both ActiveModel and ActiveRecord
Attributes API was introduced in Rails 5, so
`ActiveRecord::Type.register(:time_only, Tod::TimeOfDayType)` is only
supported since Rails 5—it's not a fallback for older versions of Rails.

In Rails 6.1 we ran into an issue where a type registered in
ActiveModel's registry was not available in ActiveRecord models. This
might be a change in Rails or load order quirk. In any case code
examples in the documentation suggest that a type should be registered
in both registries.